Time: 8 ms (84.41%), Space: 44 MB (83.75%) - LeetHub

+++ 0326-power-of-three/0326-power-of-three.java
@@ -3,7 +3,7 @@ class Solution {
         if(n<=0) return false;
         while(n%3 == 0)
         {
-            n = (int)n/3;
+            n = n/3;
 
         }
         return (n==1) ? true : false;
